+++ src/main/factories/controllers/user-preference-game-place/save-user-preference-game-place-controller-factory.ts
@@ -1,4 +1,4 @@
-import { SaveUserPreferenceGamePlaceZodValidation } from '@/validators/user-preferece-game-place/save-user-preference-game-place-zod-validation'
+import { SaveUserPreferenceGamePlaceZodValidation } from '@/validators/zod/user-preferece-game-place/save-user-preference-game-place-zod-validation'
 import { makeSaveUserPreferenceGamePlaceUsecase } from '../../usecases/user-preference-game-place/save-user-preference-game-place-usecase-factory'
 import { type Controller } from '@/presentation/contracts'
 import { SaveUserPreferenceGamePlaceController } from '@/presentation/controllers/user-preference-game-place/save-user-preference-game-place/save-user-preference-game-place-controller'
