+++ src/main/factories/controllers/user-social-media/save-user-social-media-controller-factory.ts
@@ -1,5 +1,5 @@
 import { type Controller } from '@/presentation/contracts'
-import { SaveUserSocialMediaValidation } from '@/validators/user-social-medial/save-user-social-media-zod-validation'
+import { SaveUserSocialMediaValidation } from '@/validators/zod/user-social-medial/save-user-social-media-zod-validation'
 import { makeSaveUserSocialMediaUseCase } from '../../usecases/user-social-media/save-user-social-media-usecase-factory'
 import { SaveUserSocialMediaController } from '@/presentation/controllers/user-social-media/save-user-social-media/save-user-social-media-controller'
