+++ src/main/factories/controllers/user/add-user-preference-controller-factory.ts
@@ -1,6 +1,6 @@
 import { type Controller } from '@/presentation/contracts'
 import { AddUserPreferenceController } from '@/presentation/controllers/user/add-user-preference/add-user-preference-controller'
-import { AddUserPreferenceZodValidation } from '@/validators/user/add-user-preference/add-user-preference-zod-validation'
+import { AddUserPreferenceZodValidation } from '@/validators/zod/user/add-user-preference/add-user-preference-zod-validation'
 import { makeAddUserPreferenceUseCase } from '../../usecases/user/add-user-preference-usecase-factory'
 
 export const makeAddUserPreferenceController = (): Controller => {
