+++ src/main/factories/controllers/user/signup-controller-factory.ts
@@ -1,6 +1,6 @@
 import { makeAddUserUseCase } from '@/main/factories/usecases/user/add-user-usecase-factory'
 import { SignUpController } from '@/presentation/controllers/user/signup/signup-controller'
-import { SignUpZodValidation } from '@/validators/user/signup/signup-zod-validation'
+import { SignUpZodValidation } from '@/validators/zod/user/signup/signup-zod-validation'
 import { type Controller } from '@/presentation/contracts/controller'
 
 export const makeSignUpController = (): Controller => {
