+++ src/main/factories/controllers/user/update-user-controller-factory.ts
@@ -1,6 +1,6 @@
 import { type Controller } from '@/presentation/contracts/controller'
 import { UpdateUserController } from '@/presentation/controllers/user/update-user/update-user-controller'
-import { UpdateUserZodValidation } from '@/validators/user/update-user/update-user-zod-validation'
+import { UpdateUserZodValidation } from '@/validators/zod/user/update-user/update-user-zod-validation'
 import { makeUpdateUserUseCase } from '../../usecases/user/update-user-usecase-factory'
 
 export const makeUpdateUserController = (): Controller => {
