+++ src/validators/helpers/zod-helper.ts
@@ -1,12 +1,8 @@
 import { left, right, type Either } from '@/shared/either'
-import { ZodError, type ZodSchema } from 'zod'
-import { fromZodError } from 'zod-validation-error'
 import { ValidationError } from '@/validators/errors/validation-error'
-
-export type ZodHelperData = {
-  value: any
-  schema: ZodSchema
-}
+import { ZodError } from 'zod'
+import { fromZodError } from 'zod-validation-error'
+import { type ZodHelperData } from '../types/zod-helper-data'
 
 export class ZodHelper {
   static check (data: ZodHelperData): Either<Error, null> {
