+++ src/validators/types/zod-helper-data.ts
@@ -1,0 +1,6 @@
+import { type ZodSchema } from 'zod'
+
+export type ZodHelperData = {
+  value: any
+  schema: ZodSchema
+}
